Switch to custom math library
Want to use floats rather than doubles to reduce the memory used by a STL
file in half.  This is also a more appropriate data type because it
avoids a translation in data types when writing a STL binary file and
when rendering with OpenGL.

Having full control over the math library allows for better integration
throughout the program.  Copy operations may be avoided when reading a
STL ASCII file that previously conformed to a rigid interface.  Vertices
may be moved a certain distance and the distance between them may be
calculated in both 2D and 3D.

Model how operators were overloaded with MathNet.Spatial.Euclidean since
this were very handy means of expression.

+++ src/Geometry/Polygon3D.cs
@@ -1,5 +1,5 @@
 using System.Collections.Generic;
-using MathNet.Spatial.Euclidean;
+using CalcNet.Spatial.Euclidean;
 
 namespace ScaffoldingGenerator.Geometry
 {
@@ -14,10 +14,10 @@ namespace ScaffoldingGenerator.Geometry
         {
             Facets = facets;
 
-            double minX, minY, minZ;
-            double maxX, maxY, maxZ;
-            minX = minY = minZ = double.MaxValue;
-            maxX = maxY = maxZ = double.MinValue;
+            float minX, minY, minZ;
+            float maxX, maxY, maxZ;
+            minX = minY = minZ = float.MaxValue;
+            maxX = maxY = maxZ = float.MinValue;
             foreach (Facet facet in facets)
             {
                 foreach (Point3D vertex in facet.Vertices)
